Splitting out callback methods.
NativeCrypto.cpp is huge and hard to reason about. This is a start at
splitting out a few related pieces.
